+++ src/domains/components/NewSlaveZone.js
