+++ test/domains/components/EditNSRecord.spec.js
@@ -5,9 +5,9 @@ import { expect } from 'chai';
 
 import { api } from '@/data';
 import { expectRequest } from '@/common';
-import EditSRVRecord from '~/dnsmanager/components/EditSRVRecord';
+import EditNSRecord from '~/domains/components/EditNSRecord';
 
-describe('dnsmanager/components/EditSRVRecord', () => {
+describe('domains/components/EditNSRecord', () => {
   const sandbox = sinon.sandbox.create();
 
   afterEach(() => {
@@ -16,11 +16,11 @@ describe('dnsmanager/components/EditSRVRecord', () => {
 
   const dispatch = sandbox.stub();
 
-  it('renders fields correctly for SRV record', () => {
+  it('renders fields correctly', () => {
     const currentZone = api.dnszones.dnszones['1'];
-    const currentRecord = currentZone._records.records[6];
+    const currentRecord = currentZone._records.records[4];
     const page = mount(
-      <EditSRVRecord
+      <EditNSRecord
         dispatch={dispatch}
         zone={currentZone}
         id={currentRecord.id}
@@ -28,34 +28,22 @@ describe('dnsmanager/components/EditSRVRecord', () => {
       />
     );
 
-    const service = page.find('#service');
-    expect(service.props().value).to.equal(currentRecord.service);
+    const nameserver = page.find('#nameserver');
+    expect(nameserver.props().value).to.equal(currentRecord.target);
 
-    const protocol = page.find('#protocol');
-    expect(protocol.props().value).to.equal(currentRecord.protocol);
-
-    const target = page.find('#target');
-    expect(target.props().value).to.equal(currentRecord.target);
-
-    const priority = page.find('#priority');
-    expect(priority.props().value).to.equal(currentRecord.priority);
-
-    const weight = page.find('#weight');
-    expect(weight.props().value).to.equal(currentRecord.weight);
-
-    const port = page.find('#port');
-    expect(port.props().value).to.equal(currentRecord.port);
+    const subdomain = page.find('#subdomain');
+    expect(subdomain.props().value).to.equal(currentRecord.name || currentZone.dnszone);
 
     const ttl = page.find('#ttl');
     expect(+ttl.props().value).to.equal(currentRecord.ttl_sec || currentZone.ttl_sec);
   });
 
-  it('submits data onsubmit and closes modal for SRV record', async () => {
+  it('submits data onsubmit and closes modal', async () => {
     const currentZone = api.dnszones.dnszones['1'];
-    const currentRecord = currentZone._records.records[6];
+    const currentRecord = currentZone._records.records[4];
     const close = sandbox.spy();
     const page = mount(
-      <EditSRVRecord
+      <EditNSRecord
         dispatch={dispatch}
         zone={currentZone}
         id={currentRecord.id}
@@ -66,12 +54,8 @@ describe('dnsmanager/components/EditSRVRecord', () => {
     const changeInput = (name, value) =>
       page.instance().setState({ [name]: value });
 
-    changeInput('service', '_ips');
-    changeInput('protocol', '_udp');
-    changeInput('target', 'ns2.service.com');
-    changeInput('priority', 77);
-    changeInput('weight', 7);
-    changeInput('port', 777);
+    changeInput('nameserver', 'ns1.tester1234.com');
+    changeInput('subdomain', 'tester1234.com');
     changeInput('ttl', 3600);
 
     await page.find('Form').props().onSubmit();
@@ -83,23 +67,19 @@ describe('dnsmanager/components/EditSRVRecord', () => {
     await expectRequest(fn, `/dns/zones/${currentZone.id}/records/${currentRecord.id}`, {
       method: 'PUT',
       body: {
-        service: '_ips',
-        protocol: '_udp',
-        target: 'ns2.service.com',
-        priority: 77,
-        weight: 7,
-        port: 777,
+        target: 'ns1.tester1234.com',
+        name: 'tester1234.com',
         ttl_sec: 3600,
-        type: 'SRV',
+        type: 'NS',
       },
     });
   });
 
-  it('creates a new SRV record and closes the modal', async () => {
+  it('creates a new NS record and closes the modal', async () => {
     const currentZone = api.dnszones.dnszones['1'];
     const close = sandbox.spy();
     const page = mount(
-      <EditSRVRecord
+      <EditNSRecord
         dispatch={dispatch}
         zone={currentZone}
         close={close}
@@ -109,12 +89,8 @@ describe('dnsmanager/components/EditSRVRecord', () => {
     const changeInput = (name, value) =>
       page.instance().setState({ [name]: value });
 
-    changeInput('service', '_ips');
-    changeInput('protocol', '_udp');
-    changeInput('target', 'ns2.service.com');
-    changeInput('priority', 77);
-    changeInput('weight', 7);
-    changeInput('port', 777);
+    changeInput('nameserver', 'ns1.tester1234.com');
+    changeInput('subdomain', 'tester1234.com');
     changeInput('ttl', 3600);
 
     dispatch.reset();
@@ -127,14 +103,10 @@ describe('dnsmanager/components/EditSRVRecord', () => {
     await expectRequest(fn, `/dns/zones/${currentZone.id}/records/`, {
       method: 'POST',
       body: {
-        service: '_ips',
-        protocol: '_udp',
-        target: 'ns2.service.com',
-        priority: 77,
-        weight: 7,
-        port: 777,
+        target: 'ns1.tester1234.com',
+        name: 'tester1234.com',
         ttl_sec: 3600,
-        type: 'SRV',
+        type: 'NS',
       },
     });
   });
