+++ test/domains/components/EditSRVRecord.spec.js
@@ -5,9 +5,9 @@ import { expect } from 'chai';
 
 import { api } from '@/data';
 import { expectRequest } from '@/common';
-import EditTXTRecord from '~/dnsmanager/components/EditTXTRecord';
+import EditSRVRecord from '~/domains/components/EditSRVRecord';
 
-describe('dnsmanager/components/EditTXTRecord', () => {
+describe('domains/components/EditSRVRecord', () => {
   const sandbox = sinon.sandbox.create();
 
   afterEach(() => {
@@ -16,11 +16,11 @@ describe('dnsmanager/components/EditTXTRecord', () => {
 
   const dispatch = sandbox.stub();
 
-  it('renders fields correctly for TXT record', () => {
+  it('renders fields correctly for SRV record', () => {
     const currentZone = api.dnszones.dnszones['1'];
-    const currentRecord = currentZone._records.records[4];
+    const currentRecord = currentZone._records.records[6];
     const page = mount(
-      <EditTXTRecord
+      <EditSRVRecord
         dispatch={dispatch}
         zone={currentZone}
         id={currentRecord.id}
@@ -28,22 +28,34 @@ describe('dnsmanager/components/EditTXTRecord', () => {
       />
     );
 
-    const textname = page.find('#textname');
-    expect(textname.props().value).to.equal(currentRecord.name);
+    const service = page.find('#service');
+    expect(service.props().value).to.equal(currentRecord.service);
 
-    const textvalue = page.find('#textvalue');
-    expect(textvalue.props().value).to.equal(currentRecord.target);
+    const protocol = page.find('#protocol');
+    expect(protocol.props().value).to.equal(currentRecord.protocol);
+
+    const target = page.find('#target');
+    expect(target.props().value).to.equal(currentRecord.target);
+
+    const priority = page.find('#priority');
+    expect(priority.props().value).to.equal(currentRecord.priority);
+
+    const weight = page.find('#weight');
+    expect(weight.props().value).to.equal(currentRecord.weight);
+
+    const port = page.find('#port');
+    expect(port.props().value).to.equal(currentRecord.port);
 
     const ttl = page.find('#ttl');
     expect(+ttl.props().value).to.equal(currentRecord.ttl_sec || currentZone.ttl_sec);
   });
 
-  it('submits data onsubmit and closes modal for TXT record', async () => {
+  it('submits data onsubmit and closes modal for SRV record', async () => {
     const currentZone = api.dnszones.dnszones['1'];
-    const currentRecord = currentZone._records.records[4];
+    const currentRecord = currentZone._records.records[6];
     const close = sandbox.spy();
     const page = mount(
-      <EditTXTRecord
+      <EditSRVRecord
         dispatch={dispatch}
         zone={currentZone}
         id={currentRecord.id}
@@ -54,8 +66,12 @@ describe('dnsmanager/components/EditTXTRecord', () => {
     const changeInput = (name, value) =>
       page.instance().setState({ [name]: value });
 
-    changeInput('textname', 'somename');
-    changeInput('textvalue', 'someval');
+    changeInput('service', '_ips');
+    changeInput('protocol', '_udp');
+    changeInput('target', 'ns2.service.com');
+    changeInput('priority', 77);
+    changeInput('weight', 7);
+    changeInput('port', 777);
     changeInput('ttl', 3600);
 
     await page.find('Form').props().onSubmit();
@@ -67,19 +83,23 @@ describe('dnsmanager/components/EditTXTRecord', () => {
     await expectRequest(fn, `/dns/zones/${currentZone.id}/records/${currentRecord.id}`, {
       method: 'PUT',
       body: {
-        target: 'someval',
-        name: 'somename',
+        service: '_ips',
+        protocol: '_udp',
+        target: 'ns2.service.com',
+        priority: 77,
+        weight: 7,
+        port: 777,
         ttl_sec: 3600,
-        type: 'TXT',
+        type: 'SRV',
       },
     });
   });
 
-  it('creates a new TXT record and closes the modal', async () => {
+  it('creates a new SRV record and closes the modal', async () => {
     const currentZone = api.dnszones.dnszones['1'];
     const close = sandbox.spy();
     const page = mount(
-      <EditTXTRecord
+      <EditSRVRecord
         dispatch={dispatch}
         zone={currentZone}
         close={close}
@@ -89,8 +109,12 @@ describe('dnsmanager/components/EditTXTRecord', () => {
     const changeInput = (name, value) =>
       page.instance().setState({ [name]: value });
 
-    changeInput('textname', 'somename');
-    changeInput('textvalue', 'someval');
+    changeInput('service', '_ips');
+    changeInput('protocol', '_udp');
+    changeInput('target', 'ns2.service.com');
+    changeInput('priority', 77);
+    changeInput('weight', 7);
+    changeInput('port', 777);
     changeInput('ttl', 3600);
 
     dispatch.reset();
@@ -103,10 +127,14 @@ describe('dnsmanager/components/EditTXTRecord', () => {
     await expectRequest(fn, `/dns/zones/${currentZone.id}/records/`, {
       method: 'POST',
       body: {
-        target: 'someval',
-        name: 'somename',
+        service: '_ips',
+        protocol: '_udp',
+        target: 'ns2.service.com',
+        priority: 77,
+        weight: 7,
+        port: 777,
         ttl_sec: 3600,
-        type: 'TXT',
+        type: 'SRV',
       },
     });
   });
