+++ test/domains/components/NewSlaveZone.spec.js
